Reworked packet handling, add protocols, add Packet attribute with Id and PacketMap

+++ ConduitNet/Packets/Packet.cs
@@ -5,6 +5,6 @@ namespace Conduit.Net.Packets
     public class Packet
     {
         [VarInt] public int Length;
-        [VarInt] public int Id;
+        [VarInt] public byte Id;
     }
 }
